[Refactor] Move union iterator into storage primitive
