+++ be/src/storage/CMakeLists.txt
@@ -24,6 +24,7 @@ ADD_BE_LIB(StoragePrimitive
     primitive/empty_iterator.cpp
     primitive/projection_iterator.cpp
     primitive/range.cpp
+    primitive/union_iterator.cpp
 )
 
 target_link_libraries(StoragePrimitive PUBLIC
@@ -179,7 +180,6 @@ set(STORAGE_FILES
     lake_meta_reader.cpp
     olap_meta_reader.cpp
     seek_tuple.cpp
-    union_iterator.cpp
     unique_iterator.cpp
     column_aggregate_func.cpp
     chunk_aggregator.cpp
