+++ be/src/storage/primitive/union_iterator.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "storage/union_iterator.h"
+#include "storage/primitive/union_iterator.h"
 
 #include <memory>
