+++ be/test/storage/rowset_update_state_test.cpp
@@ -31,6 +31,7 @@
 #include "storage/chunk_helper.h"
 #include "storage/olap_common.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/union_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/storage_engine.h"
@@ -38,7 +39,6 @@
 #include "storage/tablet_reader.h"
 #include "storage/tablet_reader_params.h"
 #include "storage/tablet_schema.h"
-#include "storage/union_iterator.h"
 #include "storage/update_manager.h"
 
 namespace starrocks {
