+++ be/test/storage/table_reader_remote_test.cpp
@@ -27,6 +27,7 @@
 #include "gutil/walltime.h"
 #include "runtime/descriptor_helper.h"
 #include "storage/chunk_helper.h"
+#include "storage/primitive/union_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/rowset_writer.h"
@@ -37,7 +38,6 @@
 #include "storage/table_reader.h"
 #include "storage/tablet.h"
 #include "storage/tablet_manager.h"
-#include "storage/union_iterator.h"
 #include "storage/update_manager.h"
 
 namespace starrocks {
