+++ be/test/storage/tablet_updates_test.h
@@ -35,6 +35,7 @@
 #include "storage/kv_store.h"
 #include "storage/primary_key_encoder.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/union_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/rowset_writer.h"
@@ -48,7 +49,6 @@
 #include "storage/tablet_meta_manager.h"
 #include "storage/tablet_reader.h"
 #include "storage/tablet_updates.h"
-#include "storage/union_iterator.h"
 #include "storage/update_manager.h"
 
 namespace starrocks {
